Server/requirements.txt
flask
numpy
pandas
python-telegram-bot
